adapt new classes to moved API
